# src/main/resources/warehouse-openapi.yaml
openapi: 3.0.3
info:
  title: Warehouse API
  version: 1.0.0
paths:
  /warehouse:
    get:
      summary: Get all warehouses
      operationId: getAllWarehouses
      responses:
        '200':
          description: A list of warehouses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Warehouse'
    post:
      summary: Create a new warehouse
      operationId: createWarehouse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Warehouse'
      responses:
        '201':
          description: Warehouse created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Warehouse'
        '400':
          description: Invalid input

  /warehouse/{businessUnitCode}:
    get:
      summary: Get a warehouse by its business unit code
      operationId: getWarehouseByBusinessUnitCode
      parameters:
        - name: businessUnitCode
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The requested warehouse
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Warehouse'
        '404':
          description: Warehouse not found

  /warehouse/{businessUnitCode}/archive:
    post:
      summary: Archive a warehouse
      operationId: archiveAWarehouseUnitByID
      parameters:
        - name: businessUnitCode
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Warehouse archived
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Warehouse'
        '404':
          description: Warehouse not found

  /warehouse/search:
    get:
      summary: Search and filter warehouses
      operationId: searchWarehouses
      parameters:
        - name: location
          in: query
          schema:
            type: string
        - name: minCapacity
          in: query
          schema:
            type: integer
        - name: maxCapacity
          in: query
          schema:
            type: integer
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [createdAt, capacity]
            default: createdAt
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: A paginated list of warehouses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Warehouse'

components:
  schemas:
    Warehouse:
      type: object
      properties:
        businessUnitCode:
          type: string
        location:
          type: string
        capacity:
          type: integer
        stock:
          type: integer
        createdAt:
          type: string
          format: date-time
        archivedAt:
          type: string
          format: date-time